Finished implementation of Modular Control Loop Checker. Testing and bug-fixing still to go.
Added the ability to use the modular control loop checker from the IDE menu.

Resolved conflicts due to @unused tags. Completely reverted the changes

The modular control loop checker doesn't work: It fails when trying to set a satisfied (or failed) result, as there may be a conflict between the ControlLoopChecker interface and the AbstractModelVerifier class.