Solved a few more bugs.
Currently, the selection process for selecting control loops is very bad, and means many compositions need to be made, making a very, very large automata at the end if it had a control loop. As a result, one test runs out of heap space at the end.

The storage of automata are temporarily being made as a list, and not a set, for debugging purposes. This has highlighted another problem with the selection of non-loop events.