Added functionality for selecting the merge criterion for Modular Control Loop Checker
Fixed serious bug that allowed the MCLC to merge two automata even if they shared the same counter example. This bug-fix means that the MCLC passes all tests

There may still be an outstanding bug, as bmw_fh solves in less than 0.4 seconds. Either that, or I have the wrong system which VALID fails in...