104e is done i think.
The only difference is the ordering of Nodes created, while I have C_2 as the end node, and example file has C_3.

This is because I create a End node as soon as i see a do statement,so that the end node is created earlier than other statements.

I think this is acceptable because a Do statement means there is a end state no matter it has "break" or not.

hmm, don't know how to modify the ordering yet.