MODIFY. State count of dropped trivial automata.
When a trivial automaton is about to be dropped,
TRCompositionalStateCounter will first store its state count, and only
then can the automaton be dropped. At the end of the state counting
process, the total state count of the system would be the product of the
total from the remaining transition relations and the total of the
dropped automata.