MODIFY. Disable automaton variable generation in EFSM compiler.
Automaton variables were generated to allow for guard expressions that
reference the current state of other automata. However, this feature is
not yet fully implemented, and the generation of variables causes type
errors when state names are not identifiers. For now these variables are
disabled completely, at least until they are fully implemented.