ADD. Global font and display scaling.
The IDE now adjusts its font sizes and the zoom factor based on the icon
set (Configure -> Options -> gui.editor -> Icon Set and font scaling).
The Supremica and 16x16 icon sets are not scaled, the 20x20 icon set is
scaled by 20/16 = 125%, and the 24x24 icon set is scaled by 24/16=150%.

Scaling affects (almost) all fonts as well as the editor and simulator
windows. If the user chooses another zoom factor from the drop box, it
is multiplied to the default.

The font scaling is OS and Look&Feel dependent, and should work for the
default (Metal) and Windows look&feel. I have disabled all other
options. The user can now only select (Configure -> Options -> general
-> Java look&feel) the Default look&feel Metal or the System look&feel,
which is Windows for Windows or GTK (not recommended) for Linux. I have
not tested for Mac.

closes #74