MODIFY. Copy and paste components with events.
When an automaton (or variable or instance) is copied to the clipboard,
its events are also included in the transferable. When pasting, events
found in the clipboard are added to the target module unless it already
contains events with the same name. This makes it possible to transfer
simple automata between modules without having to copy the events
manually. However, aliases and variables are not copied, nor are named
constants referenced by the automaton or its events.

closes #72