BUG FIX. Prevent selection of muliple node labels without nodes.
If the user selects or drag-selects node labels and other items in a graph, then the nodes associated with any selected node labels are automatically added to the selection. This makes it easier to select nodes and prevents exceptions when trying to copy a selection with orphaned labels to the clipboard.