MODIFY. Upgrade logging system to use log4j2.
The default logging configuration has been changed so that errors,
warnings, and info messages are displayed in the IDE's log panel. This
default can be changed through the options dialog or through the popup
menu of the log panel. Logging output can also be sent to the console
(stderr) or a log file.

All Supremica's logging interfaces have been removed and replaced by
direct calls to log4j2. The correct way to generate a log message is:
  Logger logger = LogManager.getLogger()
  logger.info("This message shall be logged.");
This will display the message in the log panel and/or elsewhere
depending on the configuration. Instead of info(), we can use methods
such as error() or warn() for different severity levels. We can also use
formatting, for example:
  logger.warn("The name {} is not recommended.", name);
For more information on the use of log4j2, please refer to its
documentation:
  https://logging.apache.org/log4j/2.0/manual/api.html

These changes require 'ant intialize' to get Supremica to run again. The
class path must include the libraries lib/log4j-api-2.9.0.jar and
lib/log4j-core-2.9.0.jar, which cannot be included in SupremicaLib.jar.
They are added to the manifest by 'ant initialize', and the Eclipse (but
not Netbeans) class path has been updated accordingly.

closes #81