MODIFY. Compiler automatically generates missing namespaces.
When compiling EFSMs, the compiler will generate qualified identifiers
to represent instances of an event with different variable values, such
as e.{x==1} and e.{x==2}. When back-importing the resulting product DES,
this would result in a module that does not compile because of a missing
namespace (e in the above example). This behaviour has been changed by
making the compiler automatically create all missing namespaces. All
known compiler tests now can back-import and compile to the same module
as before.