ADD. Include prioritised events in legacy check.
If the user opens an XML file that cannot be converted to Waters without loss of information, they are offered to edit it (losing information) or to open in the Analyzer only. Previously,  only projects with an animation were considered as incompatible, now this also includes projects that have an event marked as not prioritised.

closes #94