MODIFY. Scale Windows fonts even if at 100%.
In Windows, some fonts may be scaled through the control panel. This conflicts with Supremica's scaling and results in incorrect proportions of icons and fonts. This change overrides the control panel in all cases. The user will have to choose the appropriate icon set (Configure -> Options -> gui.editor -> Icon set) to get the correct scaling.