Refactor the guard generation procedure for the monolithic approach
Guards are generated according to (synchronized) edges. For instance, when
edge e of automaton a is synchronized with e' (labelled with the same event)
of automaton a', the generated guard is only attached to e or e' (simple
sorting are used to make it deterministic) but all of them. Not only this is
for simplicity, but also, when user only needs to modify the guard once.

The generated guards, when attached to the model, no longer display in a
different color in order to not mess up when dumping the model (see issue #59).
The implementation always attach guards to the edge with zero or minimal
existing guards.

The new guard generation works only for the monolithic approach. Support
for the partitioning approach is on its way.