Fix the issue for variables with suffix "_curr".
Variables with suffix "_curr" are allowed to manually define. However, if
a variable is named after var_curr with var being the name of an automaton,
it will be treated as the automaton variable to keep track of the locations
of the automaton. A warning will be issued if an instance in the range of
the variable is not a location name.