MODIFY. Counterexamples can contain more than one trace.
Model verifiers now return CounterExampleProxy instead of TraceProxy in
their VerificationResult. Each CounterExampleProxy includes the
information about the referenced product DES and automata, the comment,
and one or more traces of type TraceProxy. CounterExampleProxy has
different subtypes for different properties (SafetyCounterExampleProxy,
ConflictCounterExampleProxy, LoopCounterExampleProxy). Each TraceProxy
object contains a sequence of events, possibly with state information,
and a loop index if it is a cycle. A loop index of -1 indicates that
there is no cycle, so TraceProxy has no further subtypes.