ADD. Option to disable background compiler.
The automatic compilation can be disabled by deselecting "Compile automatically while editing" in the "gui" tab of the Options dialog. If the background compiler is disabled, modules do not get compiled automatically when loaded or edited. The compiler is still started when the user switches tabs or requests verification, in which case a dialog is displayed which has a button to abort the compilation. Aborting may take some time for large modules, however.
closes #116