ADD. SAX-based unmarshaller for modules.
This is the first step towards a new marshalling and unmarshalling
system, which will make it possible to read and write XML file without
the need for the Java Architecture for XML Binding (JAXB). When
completed, the XJC task to generate Java classes from XML schemas will
no longer be needed, hopefully making it possible to compile and run
with Java versions 11 and higher. But the XJC task is replaced by a
hand-coded Java program to extract keywords from the XML schemas.

'ant initialize' is needed to recompile after this update. It will
continue to be used even after removal of the XJC task.