BUG FIX. Mapping variables back to enum values after BDD synthesis.
Changing ExtendedAutomataIndexMap such that all variables have their own mapping between the integer representation of a value and its identifier representation. For enums we now have a mapping {enumidentifier1=0, enumidentifier2=1,...} and vice versa. Since this was all done in one mapping shared between all variables, the global lower and upper bound of all variables are kept track of separately to use for BDD creation.

The above mentioned variable specific mapping is used in BDDExtendedAutomata when converting a BDD to automaton states. Instead of mapping enum variable values to an integer, now the value is correctly mapped to the enum value identifier.

BDDAbstractManager needed a change such that the new variableRange is used instead of the old variable mapping. At this point in BDDAbstractManager the variable name is not available, so the old method of checking bounds is not applicable.

BDDExtendedGuardGenerator needed a change since enum variables probably should not be sorted, and the relational operators applied to enums should be undefined.

#127