ADD. Implementation of set representation in toString.
Instead of returning a string representation of the array keeping the BDDs when the bit vector does not represent a constant number, the toString method now goes through all possible numbers the bit vector can represent in theory and checks whether the BDDs can satisfy such a number.

Incrementing a mask and comparing with the bit vector to identify all integers it can assume. Stops when the mask wraps around, which is detected by checking whether the increment operation overflows.

#125