BUG FIX. Cleaning the BDD factory if exceptions are thrown
BDDExtendedAutomata makes sure to free the BDD factory if an exception is thrown during object instantiation. Otherwise there is a risk that the reference to the factory is lost if initialize() throws an exception. If the reference is lost there is no way to free it. If the factory is not freed, new attempts to instantiate it will throw an exception.
EditorSynthesizerAction makes sure to free the BDD factory if an exception is thrown during synthesis. As above, there is otherwise a risk that the reference to factory is lost.
PCGVariableSorter rethrows exceptions instead of printing it since higher level functionality in Supremica can print it better.
#134