MODIFY. BDD synthesis can handle named constants in guards
BDDAbstractManager differentiates simple identifiers into variables and named constants. The BDDExtendedAutomata is queried for the BDD corresponding to the variable name. A named constant is replaced by its expression and a recursive call is made to convert it to BDD.
ExtendedAutomata adds three getters to its API such that the manager can query variables and named constants.
#134