REFACTOR. Using visitor pattern to parse guard/action expressions into BDDs
Creating inner class Expression2BDDBitVectorVisitor for parsing of expressions. Now the checks for 'instanceof' have been removed and replaced with different method calls depending on expression type. All original code is copied as is, except for the recursive calls to expr2BDDBitVec where the calls are instead made to new instances of the visitor class.