MODIFY. Launch scripts for Windows to support OpenJDK.
The process to locate Java under Windows now is as follows:
1. Check the JAVA_HOME environment variable.
2. Check whether a java executable is in the system path.
3. Search the registry for Java version 8.
4. Search the registry for higher Java versions.
The first of these checks that returns a Java version 8 or higher is
used.
The Visual Basic Script waters.vbs searches the registry for both
OpenJDK and Oracle versions of Java, while the Batch script waters.bat
only looks for Oracle. The latter should be sufficient, as OpenJDK
installers (unlike Oracle) seem to set JAVA_HOME and/or the system path.