REMOVE. Foreach blocks no longer have guards.
Up to now, foreach blocks (ForeachProxy) had an optional guard (WHERE
clause) to restrict the loop based on a condition. With the introduction
of conditional blocks (ConditionalProxy), this is no longer needed and
is now removed. When opening legacy modules, any guard of a foreach
block is automatically converted to a nested conditional block.