MODIFY. Avoid gradual change of window size over multiple sessions.
When the IDE first opens, it sets its preferred size based on the
properties file. However, Java may set the frame to slightly different
size. This would cause the adjusted size to be written to the properties
file, affecting the frame size in the next session. Over several
sessions, the frame size could change dramatically.
To avoid this, the initial component-resized event is now handled
differently, storing the new frame size only if vastly different from
the requested size.