ADD. Pasting into label blocks with items selected within.
If labels are pasted while a label within a label block is selected, the
insertion into the label block is done behind the selected label
(selection anchor). Except, if the selected item within the label block
is a nested block (conditional or foreach), the insertion occurs at the
end of the selected nested block.