ADD. Conditional blocks can contain EFSM variables.
In this preliminary version, the ModuleInstanceCompiler transforms any
conditionals that cannot be resolved into guard/action blocks. Future
improvements will see the reverse: compilation of all guard/action
blocks into conditionals, which will be handled by the EFSMNormaliser
and EFSMCompiler.

This will be optional, so legacy tools that rely on guard/action blocks
continue to be supported.