MODIFY. Smarter handling of foreach/conditional blocks in popups.
Graph editor popup allows for editing of a foreach and conditional block
under the mouse (instead of guard/action block).
Graph editor popup also allows for addition of foreach and conditional
block to selected label block or node.
But the options to add these items are suppressed when instantiation is
disabled through the options.