MODIFY. Throw exception when compiling group nodes with actions.
When a group node transition is overridden by a more specific transition
from a contained node (group or simple), then the negation of the guard
of the more specific transition is included as an additional guard of
the group node transition. This does not work correctly when the more
specific guard uses next-state variables (requires existential
quantification, which is not yet implemented). To avoid confusion, the
GroupNodeCompiler will now throw ActionSyntaxException when detecting
this condition.