MODIFY. More rigorous error checking in compiler.
The module instance compiler has been refactored in preparation for
output of conditionals, which is now optionally supported.
This includes more stringent checking of nested primes and assignments
in guards and conditionals. Primes are only allowed within guards and
conditionals, but not nested within array indexes, assignments, or other
primes. Assignments are only allowed as actions or within a conjunction
at the topmost level of a conditional.
The processing of false conditions also is more rigorous. If a
guard/action block or conditional within an edge's label block evaluates
to false during instantiation, its contents are still processed and any
events are added to the component's blocked events list. Errors during
this processing of labels under a false condition are suppressed.