MODIFY. Normalising compiler optimisation of blocked events.
If an event is blocked in one plant component of a product DES, it can
never occur and therefore be removed from the model without changing the
behaviour. But such removal should only occur when compiler optimisation
is enabled.
The EFSMNormaliser and EFSMCompiler have been changed to adhere to this.
With optimisation enabled, if an event is found to be disabled in all
states of a plant component, or a controllable event is found to be
disabled in all states of a specification or supervisor component, the
event is added to the blocked events list of the first component
blocking it and removed from all other components. No attempts are made
to normalise or split its guards.
With optimisation disabled, if an event is blocked in a component, it is
added to the blocked events list of that component and no guards are
generated for the event in that component. The event remains in other
components, which may lead to guard normalisation and splitting.
With or without optimisation, events blocked on an uncontrollable event
in a specification or supervisor component, or on any event in a
property component, are not considered as globally disabled. They are
added to the blocked events lists of components blocking them and
treated in the same way as without optimisation.
These corrections only apply to the normalising compiler chain. The old
compiler removes some blocked events entirely from the model, even
without optimisation. This has been left unchanged.