MODIFY. Abort supervisor reduction projection when futile.
When reducing a non-localised supervisor with projection, if the growth
limit is exceeded or an overflow occurs during determinisation, then
synthesis is aborted with an OverflowException.
For localised supervisors, the old behaviour continues, trying to reduce
the supervisor without projection and making it possible to apply
projection to some but not all supervisor components.