MODIFY. Heuristics of modular verifiers configurable as lists.
A heuristic can now be given as a sequence of enum elements
(MaxCommonEvents, EarlyNotAccept, etc.). These are entered as comma
separated string on the command line, not yet implemented in GUI. If one
heuristic fails to distinguish two components, the next one in the
sequence is used to break the tie. If only one heuristic is specified, a
default list of alternatives is appended. To suppress alternatives, a
list including the One heuristic can be used, (e.g. MaxCommonEvents,One)
as One random selection.
The Preference option has been removed, use the Plant or RealPlant
heuristics as part of the sequence instead.